+++ src/index.es6.js
@@ -1,3 +1,19 @@
+/**
+ * @license
+ * Copyright (c) 2017 CANDY LINE INC.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 'use strict';
 
 import 'source-map-support/register';
@@ -16,6 +32,10 @@ const PORT = process.env.PORT || 8100;
 const DEFAULT_PACKAGE_JSON = __dirname + '/../package.json';
 const DEFAULT_WELCOME_FLOW = __dirname + '/welcome-flow.json';
 const NODE_PALETTE_ENABLED = process.env.NODE_PALETTE_ENABLED ? process.env.NODE_PALETTE_ENABLED === 'true' : false;
+const CANDY_RED_SESSION_TIMEOUT = parseInt(process.env.CANDY_RED_SESSION_TIMEOUT || 86400);
+const CANDY_RED_ADMIN_USER_ID = process.env.CANDY_RED_ADMIN_USER_ID;
+const CANDY_RED_ADMIN_PASSWORD_ENC = process.env.CANDY_RED_ADMIN_PASSWORD_ENC;
+const CANDY_RED_LOG_LEVEL = process.env.CANDY_RED_LOG_LEVEL || 'info';
 
 export class CandyRed {
   constructor(packageJsonPath) {
@@ -25,7 +45,7 @@ export class CandyRed {
     this.server = http.createServer(this.app);
 
     // Device Management
-    this.deviceManagerStore = new DeviceManagerStore(RED);
+    this.deviceManagerStore = new DeviceManagerStore();
 
     // path to package.json
     this.packageJsonPath = packageJsonPath;
@@ -199,37 +219,13 @@ export class CandyRed {
         title: ' ** ' + longname + ' **',
         image: __dirname + '/public/images/candy-red.png'
       },
-      menu: {
-        'menu-item-help': {
-          label: 'Powered By Node-RED',
-          url: 'http://nodered.org/docs'
-        },
-        'menu-item-keyboard-shortcuts': true
-      }
-    };
-  }
-
-  _createCandyBoxEditorTheme(deviceId) {
-    let name = os.hostname();
-    let longname = name;
-    if (deviceId) {
-      name = deviceId;
-      longname = `${os.hostname()} (${name})`;
-    }
-    return {
-      page: {
-        title: 'CANDY BOX@' + name,
-        favicon: __dirname + '/public/images/candy-box.ico',
-        css: __dirname + '/public/css/candy-box.css'
-      },
-      header: {
-        title: ' ** ' + longname + ' **',
-        image: __dirname + '/public/images/candy-box.png'
+      login: {
+        image: __dirname + '/public/images/logo.png'
       },
       menu: {
         'menu-item-help': {
-          label: 'Powered By Node-RED',
-          url: 'http://nodered.org/docs'
+          label: 'Users Forum',
+          url: 'https://forums.candy-line.io/c/candy-red'
         },
         'menu-item-keyboard-shortcuts': true
       }
@@ -238,30 +234,25 @@ export class CandyRed {
 
   _inspectBoardStatus(inputPackageJsonPath) {
     return Promise.all([
-      this.deviceManagerStore.deviceState.testIfCANDYIoTInstalled(),
-      this.deviceManagerStore.deviceState.testIfLTEPi2Installed()
+      this.deviceManagerStore.deviceState.testIfCANDYBoardServiceInstalled('candy-pi-lite'),
+      this.deviceManagerStore.deviceState.testIfCANDYBoardServiceInstalled('ltepi2')
     ]).then(results => {
-      let candyIotv;
-      let ltepi2v;
+      let candyBsv;
       let deviceId;
       if (results[0][0]) {
         deviceId = results[0][0];
       }
       if (results[0][1]) {
-        candyIotv = results[0][1];
-        this.editorTheme = this._createCandyBoxEditorTheme(deviceId);
+        candyBsv = results[0][1];
       } else if (results[1][1]) {
-        ltepi2v = results[1][1];
-        this.editorTheme = this._createCandyRedEditorTheme(deviceId);
-      } else {
-        this.editorTheme = this._createCandyRedEditorTheme(deviceId);
+        candyBsv = results[1][1];
       }
+      this.editorTheme = this._createCandyRedEditorTheme(deviceId);
       this.editorTheme.palette = {
         editable: NODE_PALETTE_ENABLED
       };
       deviceId = deviceId || 'N/A';
-      candyIotv = candyIotv || 'N/A';
-      ltepi2v = ltepi2v || 'N/A';
+      candyBsv = candyBsv || 'N/A';
       return new Promise((resolve, reject) => {
         fs.stat(inputPackageJsonPath, err => {
           if (err) {
@@ -274,15 +265,14 @@ export class CandyRed {
           fs.readFile(packageJsonPath, (err, data) => {
             if (err) {
               return resolve({
-                candyIotv: candyIotv,
+                candyBsv: candyBsv,
                 candyRedv: 'N/A'
               });
             }
             let packageJson = JSON.parse(data);
             return resolve({
               deviceId: deviceId,
-              candyIotv: candyIotv,
-              ltepi2v: ltepi2v,
+              candyBsv: candyBsv,
               candyRedv: packageJson.version || 'N/A'
             });
           });
@@ -297,7 +287,7 @@ export class CandyRed {
     function exitHandler(err) {
       console.log('[CANDY RED] Bye');
       if (RED.settings && RED.settings.exitHandlers) {
-        RED.settings.exitHandlers.forEach(handler => {
+        RED.settings.exitHandlers.forEach((handler) => {
           try {
             handler(RED);
           } catch (err) {
@@ -315,19 +305,21 @@ export class CandyRed {
         process.exit(err);
       }
     }
-    process.on('exit', exitHandler);
     process.on('SIGINT', exitHandler);
     process.on('uncaughtException', exitHandler);
+    process.on('unhandledRejection', (err) => {
+      console.log('unhandledRejection', err);
+    });
   }
 
   _createREDSettigngs(versions) {
-    return {
+    let settings = {
       flowFilePretty: true,
       verbose: true,
       disableEditor: false,
       httpAdminRoot: '/red/',
       httpNodeRoot: '/api/',
-      userDir: (process.env.HOME || process.env.USERPROFILE) + '/.node-red',
+      userDir: (process.env.CANDY_RED_HOME || process.env.HOME || process.env.USERPROFILE) + '/.node-red',
       flowFile: this.flowFile,
       functionGlobalContext: {
       },
@@ -336,10 +328,29 @@ export class CandyRed {
       ],
       deviceManagerStore: this.deviceManagerStore,
       editorTheme: this.editorTheme,
-      candyIotVersion: versions.candyIotv,
+      candyBoardServiceVersion: versions.candyBsv,
       candyRedVersion: versions.candyRedv,
-      deviceId: versions.deviceId
+      deviceId: versions.deviceId,
+      logging: {
+        console: {
+          level: CANDY_RED_LOG_LEVEL,
+          metrics: false,
+          audit: false
+        }
+      }
     };
+    if (CANDY_RED_ADMIN_USER_ID && CANDY_RED_ADMIN_PASSWORD_ENC) {
+      settings.adminAuth = {
+        sessionExpiryTime: CANDY_RED_SESSION_TIMEOUT,
+        type: 'credentials',
+        users: [{
+          username: CANDY_RED_ADMIN_USER_ID,
+          password: CANDY_RED_ADMIN_PASSWORD_ENC,
+          permissions: '*'
+        }]
+      };
+    }
+    return settings;
   }
 
 }