+++ tests/device-manager.test.es6.js
@@ -1,3 +1,19 @@
+/**
+ * @license
+ * Copyright (c) 2017 CANDY LINE INC.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 'use strict';
 
 import * as sinon from 'sinon';
@@ -111,9 +127,9 @@ describe('DeviceState', () => {
     sandbox.restore();
   });
 
-  describe('#testIfCANDYIoTInstalled()', () => {
-    it('should return whether or not CANDY IoT board is installed', done => {
-      state.testIfCANDYIoTInstalled().then(version => {
+  describe('#testIfCANDYBoardServiceInstalled("candy-pi-lite")', () => {
+    it('should return whether or not CANDY Pi Lite board is installed', done => {
+      state.testIfCANDYBoardServiceInstalled('candy-pi-lite').then(version => {
         console.log(`installed version? => [${version}]`);
         done();
       }).catch(err => {
@@ -121,7 +137,7 @@ describe('DeviceState', () => {
       });
     });
 
-    it('should return the installed CANDY IoT Board version', done => {
+    it('should return the installed CANDY Pi Lite Board version', done => {
       let stubCproc = sandbox.stub(cproc);
       let systemctl = sandbox.stub({
         on: () => {}
@@ -135,24 +151,24 @@ describe('DeviceState', () => {
       stdout.on.onFirstCall().yields(JSON.stringify({
         version: '1234'
       }));
-      let ciot = sandbox.stub({
+      let candy = sandbox.stub({
         stdout: stdout,
         on: () => {}
       });
-      stubCproc.spawn.onSecondCall().returns(ciot);
-      ciot.on.onFirstCall().yields(0);
+      stubCproc.spawn.onSecondCall().returns(candy);
+      candy.on.onFirstCall().yields(0);
 
       state.deviceId = 'my:deviceId';
-      state.testIfCANDYIoTInstalled().then(version => {
+      state.testIfCANDYBoardServiceInstalled('candy-pi-lite').then(version => {
         assert.deepEqual(['my:deviceId', '1234'], version);
-        assert.isTrue(ciot.on.called);
+        assert.isTrue(candy.on.called);
         done();
       }).catch(err => {
         done(err);
       });
     });
 
-    it('should return the empty version as ciot info version command fails but the board is detected', done => {
+    it('should return the empty version as candy service version command fails but the board is dtected', done => {
       let stubCproc = sandbox.stub(cproc);
       let systemctl = sandbox.stub({
         on: () => {}
@@ -163,24 +179,24 @@ describe('DeviceState', () => {
       let stdout = sandbox.stub({
         on: () => {}
       });
-      let ciot = sandbox.stub({
+      let candy = sandbox.stub({
         stdout: stdout,
         on: () => {}
       });
-      stubCproc.spawn.withArgs('ciot', ['info', 'version'], { timeout: 1000 }).returns(ciot);
-      ciot.on.onFirstCall().yields(1);
+      stubCproc.spawn.withArgs('candy', ['service', 'version'], { timeout: 1000 }).returns(candy);
+      candy.on.onFirstCall().yields(1);
 
       state.deviceId = 'my:deviceId';
-      state.testIfCANDYIoTInstalled().then(version => {
+      state.testIfCANDYBoardServiceInstalled('candy-pi-lite').then(version => {
         assert.deepEqual(['my:deviceId', 'offline'], version);
-        assert.isTrue(ciot.on.called);
+        assert.isTrue(candy.on.called);
         done();
       }).catch(err => {
         done(err);
       });
     });
 
-    it('should return the empty version as systemctl is-enabled candy-iot fails', done => {
+    it('should return the empty version as systemctl is-enabled ltepi2 fails', done => {
       let stubCproc = sandbox.stub(cproc);
       let systemctl = sandbox.stub({
         on: () => {}
@@ -189,7 +205,7 @@ describe('DeviceState', () => {
       stubCproc.spawn.onFirstCall().returns(systemctl);
 
       state.deviceId = 'my:deviceId';
-      state.testIfCANDYIoTInstalled().then(version => {
+      state.testIfCANDYBoardServiceInstalled('candy-pi-lite').then(version => {
         assert.deepEqual(['my:deviceId', null], version);
         assert.isTrue(systemctl.on.called);
         done();
@@ -197,12 +213,11 @@ describe('DeviceState', () => {
         done(err);
       });
     });
-
   });
 
-  describe('#testIfLTEPi2Installed()', () => {
+  describe('#testIfCANDYBoardServiceInstalled("ltepi2")', () => {
     it('should return whether or not LTEPi2 board is installed', done => {
-      state.testIfLTEPi2Installed().then(version => {
+      state.testIfCANDYBoardServiceInstalled('ltepi2').then(version => {
         console.log(`installed version? => [${version}]`);
         done();
       }).catch(err => {
@@ -232,7 +247,7 @@ describe('DeviceState', () => {
       candy.on.onFirstCall().yields(0);
 
       state.deviceId = 'my:deviceId';
-      state.testIfLTEPi2Installed().then(version => {
+      state.testIfCANDYBoardServiceInstalled('ltepi2').then(version => {
         assert.deepEqual(['my:deviceId', '1234'], version);
         assert.isTrue(candy.on.called);
         done();
@@ -260,7 +275,7 @@ describe('DeviceState', () => {
       candy.on.onFirstCall().yields(1);
 
       state.deviceId = 'my:deviceId';
-      state.testIfLTEPi2Installed().then(version => {
+      state.testIfCANDYBoardServiceInstalled('ltepi2').then(version => {
         assert.deepEqual(['my:deviceId', 'offline'], version);
         assert.isTrue(candy.on.called);
         done();
@@ -278,7 +293,7 @@ describe('DeviceState', () => {
       stubCproc.spawn.onFirstCall().returns(systemctl);
 
       state.deviceId = 'my:deviceId';
-      state.testIfCANDYIoTInstalled().then(version => {
+      state.testIfCANDYBoardServiceInstalled('ltepi2').then(version => {
         assert.deepEqual(['my:deviceId', null], version);
         assert.isTrue(systemctl.on.called);
         done();
@@ -380,7 +395,7 @@ describe('DeviceManager', () => {
       stubCproc.spawn.onFirstCall().returns(ciot);
       ciot.on.yields();
 
-      manager.ciotSupported = true;
+      manager.candyBoardServiceSupported = true;
       manager._performInspect({
         cat: 'sys',
         act: 'inspect'
@@ -502,57 +517,6 @@ describe('DeviceManagerStore', () => {
         done(err);
       });
     });
-  });
+  }); /* #_onFlowFileRemovedFunc() */
 
-  describe('#initWsClient()', () => {
-    it('should initialize a websocket client as a primary device manager client', done => {
-      let account = {};
-      let listenerConfig = sandbox.stub({
-        registerInputNode: () => {}
-      });
-      let accountConfig = sandbox.stub({
-        accountFqn: 'TEST@localhost',
-        on: () => {}
-      });
-      let webSocketListeners = sandbox.stub({
-        get: () => {}
-      });
-      webSocketListeners.get.returns(listenerConfig);
-      assert.isFalse(store.isWsClientInitialized(accountConfig.accountFqn));
-      store.initWsClient(account, accountConfig, webSocketListeners);
-      assert.isTrue(store.isWsClientInitialized(accountConfig.accountFqn));
-      assert.isTrue(store.store[accountConfig.accountFqn].primary);
-      store.initWsClient(account, accountConfig, webSocketListeners);
-      assert.equal(1, Object.keys(store.store).length);
-      done();
-    });
-
-    it('should initialize a websocket client as NOT a primary device manager client', done => {
-      let account = {};
-      let listenerConfig = sandbox.stub({
-        registerInputNode: () => {}
-      });
-      let accountConfig = sandbox.stub({
-        accountFqn: 'TEST@localhost',
-        on: () => {}
-      });
-      let webSocketListeners = sandbox.stub({
-        get: () => {}
-      });
-      webSocketListeners.get.returns(listenerConfig);
-      store.initWsClient(account, accountConfig, webSocketListeners);
-      assert.isTrue(store.store[accountConfig.accountFqn].primary);
-
-      let accountConfig2 = sandbox.stub({
-        accountFqn: 'TEST2@localhost',
-        on: () => {}
-      });
-      assert.isFalse(store.isWsClientInitialized(accountConfig2.accountFqn));
-      store.initWsClient(account, accountConfig2, webSocketListeners);
-      assert.isTrue(store.isWsClientInitialized(accountConfig2.accountFqn));
-      assert.equal(2, Object.keys(store.store).length);
-      assert.isNotTrue(store.store[accountConfig2.accountFqn].primary);
-      done();
-    });
-  });
 });